fix(shell): use staking data hook with parameters in shell main page
