+++ libs/shell/index-page/src/lib/delegations-block/delegations-block.tsx
@@ -1,15 +1,6 @@
-import { useMemo } from 'react';
-import { DelegationResponse } from '@evmos/provider';
+import { useCallback } from 'react';
 import { useMediaQuery } from 'react-responsive';
 import { Link, useNavigate } from 'react-router-dom';
-import {
-  useAddress,
-  useStakingDelegationQuery,
-  useStakingPoolQuery,
-  // useStakingPoolQuery,
-  useStakingRewardsQuery,
-  useStakingValidatorListQuery,
-} from '@haqq/shared';
 import {
   Container,
   Heading,
@@ -20,77 +11,61 @@ import {
 import {
   ValidatorsListDesktop,
   ValidatorsListMobile,
-  // useValidatorsSort,
 } from '@haqq/staking/ui-kit';
-import {
-  sortValidatorsByToken,
-  splitValidators,
-  useValidatorsSortState,
-} from '@haqq/staking/utils';
-
-function getDelegatedValidatorsAddresses(
-  delegations: DelegationResponse[] | null | undefined,
-) {
-  if (!delegations) {
-    return [];
-  }
-
-  return delegations
-    .map((del) => {
-      if (Number.parseInt(del.balance.amount, 10) > 0) {
-        return del.delegation.validator_address;
-      }
-
-      return null;
-    })
-    .filter(Boolean);
-}
+import { useStakingData } from '@haqq/staking/utils';
 
 export function DelegationsBlock() {
-  const { haqqAddress } = useAddress();
   const {
-    data: validatorsList,
-    error,
+    totalStaked,
     status,
-  } = useStakingValidatorListQuery(1000);
-  const { data: rewardsInfo } = useStakingRewardsQuery(haqqAddress);
-  const { data: delegationInfo } = useStakingDelegationQuery(haqqAddress);
-  const { data: stakingPool } = useStakingPoolQuery();
+    error,
+    validators,
+    delegationInfo,
+    rewardsInfo,
+    sortState,
+    setSortState,
+  } = useStakingData({
+    showOnlyMyDelegation: true,
+    inactiveValidatorsVisible: false,
+  });
   const navigate = useNavigate();
   const isTablet = useMediaQuery({
     query: `(max-width: 1023px)`,
   });
 
-  const splittedValidators = useMemo(() => {
-    const { active, inactive, jailed } = splitValidators(validatorsList ?? []);
+  // const handleMobileSortChange = useCallback(
+  //   (mobileSortKey: string) => {
+  //     if (mobileSortKey === 'random') {
+  //       setSortState({
+  //         key: 'random',
+  //         direction: null,
+  //       });
+  //     }
 
-    return [
-      ...sortValidatorsByToken(active),
-      ...sortValidatorsByToken(inactive),
-      ...sortValidatorsByToken(jailed),
-    ];
-  }, [validatorsList]);
-  const valWithDelegationAddr = useMemo(() => {
-    const delegatedVals = getDelegatedValidatorsAddresses(
-      delegationInfo?.delegation_responses,
-    );
-    return delegatedVals;
-  }, [delegationInfo]);
-  const valToRender = useMemo(() => {
-    return splittedValidators
-      .filter((val) => {
-        return valWithDelegationAddr.includes(val.operator_address);
-      })
-      .filter((validator) => {
-        return validator.status === 'BOND_STATUS_BONDED';
+  //     const [key, direction] = mobileSortKey.split('-') as [
+  //       string,
+  //       SortDirection,
+  //     ];
+
+  //     setSortState({
+  //       key,
+  //       direction,
+  //     });
+  //   },
+  //   [setSortState],
+  // );
+
+  const handleDesktopSortClick = useCallback(
+    (key: string) => {
+      setSortState((state) => {
+        return {
+          key,
+          direction: state?.direction === 'asc' ? 'desc' : 'asc',
+        };
       });
-  }, [splittedValidators, valWithDelegationAddr]);
-
-  const totalStaked = useMemo(() => {
-    return Number.parseInt(stakingPool?.bonded_tokens ?? '0') / 10 ** 18;
-  }, [stakingPool?.bonded_tokens]);
-
-  const { sortState } = useValidatorsSortState();
+    },
+    [setSortState],
+  );
 
   return (
     <Container>
@@ -118,11 +93,11 @@ export function DelegationsBlock() {
 
       {status === 'success' && (
         <div>
-          {valToRender.length ? (
+          {validators.length ? (
             <div>
               {isTablet ? (
                 <ValidatorsListMobile
-                  validators={valToRender}
+                  validators={validators}
                   delegationInfo={delegationInfo}
                   rewardsInfo={rewardsInfo}
                   onValidatorClick={(validatorAddress: string) => {
@@ -132,7 +107,7 @@ export function DelegationsBlock() {
                 />
               ) : (
                 <ValidatorsListDesktop
-                  validators={valToRender}
+                  validators={validators}
                   delegationInfo={delegationInfo}
                   rewardsInfo={rewardsInfo}
                   onValidatorClick={(validatorAddress: string) => {
@@ -140,7 +115,7 @@ export function DelegationsBlock() {
                   }}
                   totalStaked={totalStaked}
                   sortState={sortState}
-                  onDesktopSortClick={console.log}
+                  onDesktopSortClick={handleDesktopSortClick}
                 />
               )}
             </div>
