+++ libs/staking/ui-kit/src/components/validator-list-item-desktop/validator-list-item-desktop.tsx
@@ -54,7 +54,6 @@ export function ValidatorListItemDesktop({
     <tr
       className="cursor-pointer border-t border-[#FFFFFF26] text-[11px] leading-[18px] transition-[background] duration-75 hover:bg-white hover:bg-opacity-[2.5%] md:text-[16px] md:leading-[26px]"
       onClick={() => {
-        // console.log(`validator/${validator.operator_address}`);
         onClick(validator.operator_address);
       }}
     >
