+++ libs/staking/utils/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/sort-validators';
 export * from './lib/split-validators';
 export * from './lib/hooks/use-validator-sort';
 export * from './lib/hooks/use-validator-filter';
+export * from './lib/hooks/use-staking-data';
